+++ pkg/server/commands/check_command.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"errors"
 	"math"
+	"sync"
 	"time"
 
+	"golang.org/x/sync/singleflight"
 	"google.golang.org/protobuf/types/known/structpb"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
@@ -26,14 +28,20 @@ const (
 )
 
 type CheckQuery struct {
-	logger          logger.Logger
-	checkResolver   graph.CheckResolver
-	typesys         *typesystem.TypeSystem
-	datastore       *storagewrappers.InstrumentedOpenFGAStorage
-	cacheController cachecontroller.CacheController
-
-	resolveNodeLimit   uint32
-	maxConcurrentReads uint32
+	logger                 logger.Logger
+	checkResolver          graph.CheckResolver
+	typesys                *typesystem.TypeSystem
+	serverCtx              context.Context
+	datastore              storage.RelationshipTupleReader
+	cacheController        cachecontroller.CacheController
+	cacheSingleflightGroup *singleflight.Group
+	cacheWaitGroup         *sync.WaitGroup
+	checkCache             storage.InMemoryCache[any]
+	checkCacheTTL          time.Duration
+	maxCheckCacheSize      uint32
+	resolveNodeLimit       uint32
+	maxConcurrentReads     uint32
+	shouldCacheIterators   bool
 }
 
 type CheckCommandParams struct {
@@ -64,21 +72,46 @@ func WithCheckCommandLogger(l logger.Logger) CheckQueryOption {
 	}
 }
 
-func WithCacheController(ctrl cachecontroller.CacheController) CheckQueryOption {
+// TODO can we make this better? There are too many caching flags.
+func WithCheckCommandCache(
+	serverCtx context.Context,
+	ctrl cachecontroller.CacheController,
+	shouldCache bool,
+	sf *singleflight.Group,
+	cc storage.InMemoryCache[any],
+	wg *sync.WaitGroup,
+	m uint32,
+	ttl time.Duration,
+) CheckQueryOption {
 	return func(c *CheckQuery) {
 		c.cacheController = ctrl
+		c.shouldCacheIterators = shouldCache
+		c.serverCtx = serverCtx
+		c.cacheSingleflightGroup = sf
+		c.cacheWaitGroup = wg
+		c.checkCache = cc
+		c.maxCheckCacheSize = m
+		c.checkCacheTTL = ttl
 	}
 }
 
-func NewCheckCommand(datastore storage.RelationshipTupleReader, checkResolver graph.CheckResolver, typesys *typesystem.TypeSystem, opts ...CheckQueryOption) *CheckQuery {
+// TODO accept CheckCommandParams so we can build the datastore object right away.
+func NewCheckCommand(
+	datastore storage.RelationshipTupleReader,
+	checkResolver graph.CheckResolver,
+	typesys *typesystem.TypeSystem,
+	opts ...CheckQueryOption,
+) *CheckQuery {
 	cmd := &CheckQuery{
-		logger:             logger.NewNoopLogger(),
-		datastore:          storagewrappers.NewInstrumentedOpenFGAStorage(datastore),
-		checkResolver:      checkResolver,
-		typesys:            typesys,
-		cacheController:    cachecontroller.NewNoopCacheController(),
-		resolveNodeLimit:   defaultResolveNodeLimit,
-		maxConcurrentReads: defaultMaxConcurrentReadsForCheck,
+		logger:               logger.NewNoopLogger(),
+		datastore:            datastore,
+		checkResolver:        checkResolver,
+		typesys:              typesys,
+		cacheController:      cachecontroller.NewNoopCacheController(),
+		resolveNodeLimit:     defaultResolveNodeLimit,
+		maxConcurrentReads:   defaultMaxConcurrentReadsForCheck,
+		shouldCacheIterators: false,
+		serverCtx:            context.TODO(),
 	}
 
 	for _, opt := range opts {
@@ -87,7 +120,10 @@ func NewCheckCommand(datastore storage.RelationshipTupleReader, checkResolver gr
 	return cmd
 }
 
-func (c *CheckQuery) Execute(ctx context.Context, params *CheckCommandParams) (*graph.ResolveCheckResponse, *graph.ResolveCheckRequestMetadata, error) {
+func (c *CheckQuery) Execute(
+	ctx context.Context,
+	params *CheckCommandParams,
+) (*graph.ResolveCheckResponse, *graph.ResolveCheckRequestMetadata, error) {
 	err := validateCheckRequest(c.typesys, params.TupleKey, params.ContextualTuples)
 	if err != nil {
 		return nil, nil, err
@@ -112,12 +148,27 @@ func (c *CheckQuery) Execute(ctx context.Context, params *CheckCommandParams) (*
 		LastCacheInvalidationTime: cacheInvalidationTime,
 	}
 
-	ctx = buildCheckContext(ctx, c.typesys, c.datastore, c.maxConcurrentReads, resolveCheckRequest.GetContextualTuples())
+	requestDatastore := storagewrappers.NewRequestStorageWrapperForCheckAPI(
+		c.serverCtx,
+		c.datastore,
+		params.ContextualTuples.GetTupleKeys(),
+		c.maxConcurrentReads,
+		c.shouldCacheIterators,
+		c.cacheSingleflightGroup,
+		c.cacheWaitGroup,
+		c.checkCache,
+		c.maxCheckCacheSize,
+		c.checkCacheTTL,
+	)
+
+	ctx = typesystem.ContextWithTypesystem(ctx, c.typesys)
+	ctx = storage.ContextWithRelationshipTupleReader(ctx, requestDatastore)
 
 	startTime := time.Now()
 	resp, err := c.checkResolver.ResolveCheck(ctx, &resolveCheckRequest)
 	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) && resolveCheckRequest.GetRequestMetadata().WasThrottled.Load() {
+		if errors.Is(err, context.DeadlineExceeded) &&
+			resolveCheckRequest.GetRequestMetadata().WasThrottled.Load() {
 			return nil, nil, &ThrottledError{Cause: err}
 		}
 
@@ -125,12 +176,16 @@ func (c *CheckQuery) Execute(ctx context.Context, params *CheckCommandParams) (*
 	}
 
 	resp.ResolutionMetadata.Duration = time.Since(startTime)
-	resp.ResolutionMetadata.DatastoreQueryCount = c.datastore.GetMetrics().DatastoreQueryCount
+	resp.ResolutionMetadata.DatastoreQueryCount = requestDatastore.GetMetrics().DatastoreQueryCount
 
 	return resp, resolveCheckRequest.GetRequestMetadata(), nil
 }
 
-func validateCheckRequest(typesys *typesystem.TypeSystem, tupleKey *openfgav1.CheckRequestTupleKey, contextualTuples *openfgav1.ContextualTupleKeys) error {
+func validateCheckRequest(
+	typesys *typesystem.TypeSystem,
+	tupleKey *openfgav1.CheckRequestTupleKey,
+	contextualTuples *openfgav1.ContextualTupleKeys,
+) error {
 	// The input tuple Key should be validated loosely.
 	if err := validation.ValidateUserObjectRelation(typesys, tuple.ConvertCheckRequestTupleKeyToTupleKey(tupleKey)); err != nil {
 		return &InvalidRelationError{Cause: err}
@@ -143,20 +198,4 @@ func validateCheckRequest(typesys *typesystem.TypeSystem, tupleKey *openfgav1.Ch
 		}
 	}
 	return nil
-}
-
-func buildCheckContext(ctx context.Context, typesys *typesystem.TypeSystem, datastore storage.RelationshipTupleReader, maxconcurrentreads uint32, contextualTuples []*openfgav1.TupleKey) context.Context {
-	ctx = typesystem.ContextWithTypesystem(ctx, typesys)
-
-	// TODO the order is wrong, see https://github.com/openfga/openfga/issues/1394
-	ctx = storage.ContextWithRelationshipTupleReader(ctx,
-		storagewrappers.NewBoundedConcurrencyTupleReader(
-			storagewrappers.NewCombinedTupleReader(
-				datastore,
-				contextualTuples,
-			),
-			maxconcurrentreads,
-		),
-	)
-	return ctx
 }